Add view model
